requirements.txt
biosppy==0.8.0
config==0.5.1
dtw==1.4.0
dtw_python==1.2.2
easydict==1.9
iterative_stratification==0.1.7
matplotlib==3.5.1
numpy==1.20.3
pandas==1.4.1
Pillow==9.2.0
pytorch_lightning==1.6.4
scikit_learn==1.1.2
scipy==1.7.3
signal_processing==0.0.5
tabulate==0.8.9
torch==1.11.0+cu113
torch_ecg==0.0.20
torchvision==0.12.0+cu113
tqdm==4.63.0
wfdb==3.2.0
